refactor: Refactor execution and utility code, update module versions
