+++ extensions/skills/version-generate/scripts/06-rules-and-edge-cases.md
@@ -14,7 +14,7 @@
 10. **追溯格式**：每條變更尾綴 `(#PR, @author) [short_hash]`，缺失欄位省略對應括號
 11. **Summary 收斂**：完整規範見 [04-output-template.md `Summary 撰寫規範`](./04-output-template.md#summary-撰寫規範強制)，含 9 條 NEVER 與 self-check checklist。常見違規（必重寫）：括號塞清單、列具體檔名／API／量化規格、中文版偷加細節、用「Adds A, B, C, D」式列舉冒充主軸。**寫完逐項勾完 self-check 才放行**，不接受「1–3 句」單一條件交差
 12. **Scope 英文唯一**：`## Scope` 段內容**一律英文**，**不附** `<details><summary>翻譯</summary>...</details>` 區塊。格式 `- \`<path>\` — <TAG1>, <TAG2>, ...`，TAG 用大寫 commit tag
-13. **不輸出 Version Mapping 表**：模板尾端**不再**附 `### Version Mapping (SemVer)` 表格與 `***` 分隔線。SemVer 規則由 frontmatter `type` 與 commit tag 自身表達，無需再列入 release note
+13. **不輸出 Version Mapping 表**：模板尾端**不再**附 `### Version Mapping (SemVer)` 表格（SemVer 規則由 frontmatter `type` 與 commit tag 自身表達，無需列入 release note）。但 `## Scope` 與結尾 `Generated by [SKILL]` footer 之間**必須**保留 `***` 分隔線
 
 ## 邊界案例
